Ensure an error is thrown on failure.
In debugging the program I found that errors were being swallowed by
promises. This commit should report the error to the console and throw
it by attaching a catch to the main promise chain.

I also added a test to examples, not sure if it's the best place to put
it, open to recommendations.

More informaton can be found here:
https://github.com/2createStudio/postcss-sprites/issues/40